Migrate test fixtures under /test to /testFixtures and added gradle java-test-fixtures plugin
