+++ sensorhub-core/src/testFixtures/java/org/sensorhub/impl/datastore/AbstractTestDataStreamStore.java
@@ -14,56 +14,53 @@ Copyright (C) 2012-2015 Sensia Software LLC. All Rights Reserved.
 
 package org.sensorhub.impl.datastore;
 
-import static org.junit.Assert.*;
-import java.io.ByteArrayOutputStream;
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import java.util.Set;
+import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
+import net.opengis.swe.v20.DataComponent;
 import org.junit.Before;
 import org.junit.Test;
-import org.sensorhub.api.command.CommandStreamInfo;
-import org.sensorhub.api.command.ICommandStreamInfo;
 import org.sensorhub.api.common.BigId;
+import org.sensorhub.api.data.DataStreamInfo;
+import org.sensorhub.api.data.IDataStreamInfo;
 import org.sensorhub.api.datastore.DataStoreException;
 import org.sensorhub.api.datastore.TemporalFilter;
-import org.sensorhub.api.datastore.command.CommandStreamFilter;
-import org.sensorhub.api.datastore.command.CommandStreamKey;
-import org.sensorhub.api.datastore.command.ICommandStreamStore;
+import org.sensorhub.api.datastore.obs.DataStreamFilter;
+import org.sensorhub.api.datastore.obs.DataStreamKey;
+import org.sensorhub.api.datastore.obs.IDataStreamStore;
 import org.sensorhub.api.feature.FeatureId;
 import org.vast.data.TextEncodingImpl;
 import org.vast.swe.SWEHelper;
 import org.vast.swe.SWEUtils;
 import org.vast.util.TimeExtent;
-import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
-import net.opengis.swe.v20.DataComponent;
+
+import java.io.ByteArrayOutputStream;
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
+import java.util.*;
+import java.util.Map.Entry;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import static org.junit.Assert.*;
 
 
 /**
  * <p>
- * Abstract base for testing implementations of ICommandStreamStore.
+ * Abstract base for testing implementations of IDataStreamStore.
  * </p>
  *
  * @author Alex Robin
  * @param <StoreType> Type of store under test
- * @since Mar 28, 2021
+ * @since Apr 14, 2018
  */
-public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandStreamStore>
+public abstract class AbstractTestDataStreamStore<StoreType extends IDataStreamStore>
 {
-    protected static int DATABASE_NUM = 5;
+    protected static int DATABASE_NUM = 2;
     protected static String PROC_UID_PREFIX = "urn:osh:test:sensor:";
     
-    protected StoreType cmdStreamStore;
-    protected Map<CommandStreamKey, ICommandStreamInfo> allCmdStreams = new LinkedHashMap<>();
-    protected Map<CommandStreamKey, ICommandStreamInfo> expectedResults = new LinkedHashMap<>();
+    protected StoreType dataStreamStore;
+    protected Map<DataStreamKey, IDataStreamInfo> allDataStreams = new LinkedHashMap<>();
+    protected Map<DataStreamKey, IDataStreamInfo> expectedResults = new LinkedHashMap<>();
     protected boolean needValidTimeAdjustment = true;
 
     protected abstract StoreType initStore() throws Exception;
@@ -73,13 +70,13 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
     @Before
     public void init() throws Exception
     {
-        this.cmdStreamStore = initStore();
+        this.dataStreamStore = initStore();
     }
     
 
-    protected CommandStreamKey addCommandStream(FeatureId sysID, DataComponent recordStruct, TimeExtent validTime) throws DataStoreException
+    protected DataStreamKey addDataStream(FeatureId sysID, DataComponent recordStruct, TimeExtent validTime) throws DataStoreException
     {
-        var builder = new CommandStreamInfo.Builder()
+        var builder = new DataStreamInfo.Builder()
             .withName(recordStruct.getName())
             .withSystem(sysID)
             .withRecordDescription(recordStruct)
@@ -89,13 +86,13 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
             builder.withValidTime(validTime);
                 
         var dsInfo = builder.build();
-        var key = cmdStreamStore.add(dsInfo);
-        allCmdStreams.put(key, dsInfo);
+        var key = dataStreamStore.add(dsInfo);
+        allDataStreams.put(key, dsInfo);
         return key;
     }
     
     
-    protected void addToExpectedResults(Entry<CommandStreamKey, ICommandStreamInfo> entry)
+    protected void addToExpectedResults(Entry<DataStreamKey, IDataStreamInfo> entry)
     {
         expectedResults.put(entry.getKey(), entry.getValue());
     }
@@ -105,13 +102,13 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
     {
         for (int idx: entryIdxList)
         {
-            var entryList = Lists.newArrayList(allCmdStreams.entrySet());
+            var entryList = Lists.newArrayList(allDataStreams.entrySet());
             addToExpectedResults(entryList.get(idx));
         }
     }
 
 
-    protected CommandStreamKey addSimpleCommandStream(FeatureId sysID, String outputName, String description, TimeExtent validTime) throws DataStoreException
+    protected DataStreamKey addSimpleDataStream(FeatureId sysID, String outputName, String description, TimeExtent validTime) throws DataStoreException
     {
         SWEHelper fac = new SWEHelper();
         var dataStruct = fac.createRecord()
@@ -124,26 +121,26 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
             .addField("txt5", fac.createText().build())
             .build();
         
-        return addCommandStream(sysID, dataStruct, validTime);
+        return addDataStream(sysID, dataStruct, validTime);
     }
     
     
-    protected CommandStreamKey addSimpleCommandStream(BigId sysID, String outputName, TimeExtent validTime) throws DataStoreException
+    protected DataStreamKey addSimpleDataStream(BigId sysID, String outputName, TimeExtent validTime) throws DataStoreException
     {
-        return addSimpleCommandStream(new FeatureId(sysID, PROC_UID_PREFIX+sysID), outputName, "command stream description", validTime);
+        return addSimpleDataStream(new FeatureId(sysID, PROC_UID_PREFIX+sysID), outputName, "datastream description", validTime);
     }
     
     
-    protected CommandStreamKey addSimpleCommandStream(BigId sysID, String outputName, String description, TimeExtent validTime) throws DataStoreException
+    protected DataStreamKey addSimpleDataStream(BigId sysID, String outputName, String description, TimeExtent validTime) throws DataStoreException
     {
-        return addSimpleCommandStream(new FeatureId(sysID, PROC_UID_PREFIX+sysID), outputName, description, validTime);
+        return addSimpleDataStream(new FeatureId(sysID, PROC_UID_PREFIX+sysID), outputName, description, validTime);
     }
 
 
-    protected void checkCommandStreamEqual(ICommandStreamInfo ds1, ICommandStreamInfo ds2)
+    protected void checkDataStreamEqual(IDataStreamInfo ds1, IDataStreamInfo ds2)
     {
         assertEquals(ds1.getSystemID(), ds2.getSystemID());
-        assertEquals(ds1.getControlInputName(), ds2.getControlInputName());
+        assertEquals(ds1.getOutputName(), ds2.getOutputName());
         assertEquals(ds1.getName(), ds2.getName());
         assertEquals(ds1.getDescription(), ds2.getDescription());
         checkDataComponentEquals(ds1.getRecordStructure(), ds2.getRecordStructure());
@@ -169,6 +166,10 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
         }        
 
         assertArrayEquals(os1.toByteArray(), os2.toByteArray());
+        
+        // also check that parent references are set properly
+        for (int i = 0; i < c2.getComponentCount(); i++)
+            assertTrue(c2.getComponent(i).getParent() == c2);
     }
     
     
@@ -178,19 +179,19 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
     }
 
 
-    protected void checkSelectedEntries(Stream<Entry<CommandStreamKey, ICommandStreamInfo>> resultStream, Map<CommandStreamKey, ICommandStreamInfo> expectedResults, CommandStreamFilter filter)
+    protected void checkSelectedEntries(Stream<Entry<DataStreamKey, IDataStreamInfo>> resultStream, Map<DataStreamKey, IDataStreamInfo> expectedResults, DataStreamFilter filter)
     {
-        System.out.println("Select command streams with " + filter);
+        System.out.println("Select datastreams with " + filter);
         
         if (needValidTimeAdjustment)
         {
             // close validTime periods when appropriate in expected results
             expectedResults = Maps.transformValues(expectedResults, v -> {
-                ICommandStreamInfo nextDs = null;
-                for (var dsInfo: allCmdStreams.values())
+                IDataStreamInfo nextDs = null;
+                for (var dsInfo: allDataStreams.values())
                 {
                     if (v.getSystemID().equals(dsInfo.getSystemID()) &&
-                        v.getControlInputName().equals(dsInfo.getControlInputName()))
+                        v.getOutputName().equals(dsInfo.getOutputName()))
                     {
                         if (v.getValidTime().endsNow() && v.getValidTime().begin().isBefore(dsInfo.getValidTime().begin()))
                         {
@@ -204,7 +205,7 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
                 if (nextDs != null)
                 {
                     var newValidTime = TimeExtent.period(v.getValidTime().begin(), nextDs.getValidTime().begin());
-                    var newDs = CommandStreamInfo.Builder.from(v).withValidTime(newValidTime);
+                    var newDs = DataStreamInfo.Builder.from(v).withValidTime(newValidTime);
                     return newDs.build();
                 }
                 
@@ -216,9 +217,9 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
     }
     
     
-    protected void checkSelectedEntries(Stream<Entry<CommandStreamKey, ICommandStreamInfo>> resultStream, Map<CommandStreamKey, ICommandStreamInfo> expectedResults)
+    protected void checkSelectedEntries(Stream<Entry<DataStreamKey, IDataStreamInfo>> resultStream, Map<DataStreamKey, IDataStreamInfo> expectedResults)
     {
-        Map<CommandStreamKey, ICommandStreamInfo> resultMap = resultStream
+        Map<DataStreamKey, IDataStreamInfo> resultMap = resultStream
                 //.peek(e -> System.out.println(e.getKey()))
                 //.peek(e -> System.out.println(Arrays.toString((double[])e.getValue().getResult().getUnderlyingObject())))
                 .collect(Collectors.toMap(e->e.getKey(), e->e.getValue()));
@@ -227,7 +228,7 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
         resultMap.forEach((k, v) -> {
             assertEquals("Invalid scope", DATABASE_NUM, k.getInternalID().getScope());
             assertTrue("Result set contains extra key "+k, expectedResults.containsKey(k));
-            checkCommandStreamEqual(expectedResults.get(k), v);
+            checkDataStreamEqual(expectedResults.get(k), v);
         });
 
         expectedResults.forEach((k, v) -> {
@@ -239,23 +240,32 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
     @Test
     public void testAddAndGetByKey() throws Exception
     {
-        // add N different command streams
+        // add N different datastreams
         var now = TimeExtent.beginAt(Instant.now());
-        int numDs = 100;
-        for (int i = 1; i < numDs; i++)
+        int numDs = 50;
+        for (int i = 1; i <= numDs; i++)
         {
             var sysID = bigId(i);
-            addSimpleCommandStream(sysID, "test1", now);
+            addSimpleDataStream(sysID, "test1", now);
         }
-        cmdStreamStore.commit();
+        dataStreamStore.commit();
         
-        // read back and check
-        forceReadBackFromStorage();
-        for (Entry<CommandStreamKey, ICommandStreamInfo> entry: allCmdStreams.entrySet())
+        // get and check
+        for (Entry<DataStreamKey, IDataStreamInfo> entry: allDataStreams.entrySet())
         {
-            ICommandStreamInfo dsInfo = cmdStreamStore.get(entry.getKey());
+            IDataStreamInfo dsInfo = dataStreamStore.get(entry.getKey());
             assertEquals(entry.getValue().getSystemID(), dsInfo.getSystemID());
-            assertEquals(entry.getValue().getControlInputName(), dsInfo.getControlInputName());
+            assertEquals(entry.getValue().getOutputName(), dsInfo.getOutputName());
+            checkDataComponentEquals(entry.getValue().getRecordStructure(), dsInfo.getRecordStructure());
+        }
+        
+        // read back and check again
+        forceReadBackFromStorage();
+        for (Entry<DataStreamKey, IDataStreamInfo> entry: allDataStreams.entrySet())
+        {
+            IDataStreamInfo dsInfo = dataStreamStore.get(entry.getKey());
+            assertEquals(entry.getValue().getSystemID(), dsInfo.getSystemID());
+            assertEquals(entry.getValue().getOutputName(), dsInfo.getOutputName());
             checkDataComponentEquals(entry.getValue().getRecordStructure(), dsInfo.getRecordStructure());
         }
     }
@@ -264,53 +274,53 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
     @Test
     public void testGetWrongKey() throws Exception
     {
-        assertNull(cmdStreamStore.get(new CommandStreamKey(0, 1L)));
-        assertNull(cmdStreamStore.get(new CommandStreamKey(0, 21L)));
+        assertNull(dataStreamStore.get(new DataStreamKey(0, 1L)));
+        assertNull(dataStreamStore.get(new DataStreamKey(0, 21L)));
         
-        // add N different command streams
+        // add N different datastreams
         var idList = new ArrayList<BigId>();
         var now = TimeExtent.beginAt(Instant.now());
         for (int i = 1; i < 5; i++)
         {
             var sysID = bigId(i);
-            var k = addSimpleCommandStream(sysID, "test1", now);
+            var k = addSimpleDataStream(sysID, "test1", now);
             idList.add(k.getInternalID());
         }
-        cmdStreamStore.commit();
+        dataStreamStore.commit();
         
-        assertNotNull(cmdStreamStore.get(new CommandStreamKey(idList.get(0))));
-        assertNull(cmdStreamStore.get(new CommandStreamKey(0, 21L)));
+        assertNotNull(dataStreamStore.get(new DataStreamKey(idList.get(0))));
+        assertNull(dataStreamStore.get(new DataStreamKey(0, 21L)));
         forceReadBackFromStorage();
-        assertNull(cmdStreamStore.get(new CommandStreamKey(0, 11L)));
-        assertNotNull(cmdStreamStore.get(new CommandStreamKey(idList.get(3))));
+        assertNull(dataStreamStore.get(new DataStreamKey(0, 11L)));
+        assertNotNull(dataStreamStore.get(new DataStreamKey(idList.get(3))));
         
     }
 
 
-    private void checkMapKeySet(Set<CommandStreamKey> keySet)
+    private void checkMapKeySet(Set<DataStreamKey> keySet)
     {
         keySet.forEach(k -> {
-            if (!allCmdStreams.containsKey(k))
+            if (!allDataStreams.containsKey(k))
                 fail("No matching key in reference list: " + k);
         });
 
-        allCmdStreams.keySet().forEach(k -> {
+        allDataStreams.keySet().forEach(k -> {
             if (!keySet.contains(k))
                 fail("No matching key in datastore: " + k);
         });
     }
 
 
-    private void checkMapValues(Collection<ICommandStreamInfo> mapValues)
+    private void checkMapValues(Collection<IDataStreamInfo> mapValues)
     {
         mapValues.forEach(ds -> {
             boolean found = false;
-            for (ICommandStreamInfo truth: allCmdStreams.values()) {
-                try { checkCommandStreamEqual(ds, truth); found = true; break; }
+            for (IDataStreamInfo truth: allDataStreams.values()) {
+                try { checkDataStreamEqual(ds, truth); found = true; break; }
                 catch (Throwable e) {}
             }
             if (!found)
-                fail("Invalid command stream: " + ds);
+                fail("Invalid datastream: " + ds);
         });
     }
 
@@ -318,103 +328,103 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
     @Test
     public void testAddAndCheckMapKeysAndValues() throws Exception
     {
-        // add N different command streams
+        // add N different datastreams
         var now = TimeExtent.beginAt(Instant.now());
         int numDs = 56;
         for (int i = numDs; i < numDs*2; i++)
         {
             var sysID = bigId(i);
-            addSimpleCommandStream(sysID, "out" + (int)(Math.random()*10), now);
+            addSimpleDataStream(sysID, "out" + (int)(Math.random()*10), now);
         }
-        cmdStreamStore.commit();
+        dataStreamStore.commit();
         
         // read back and check
         forceReadBackFromStorage();
-        checkMapKeySet(cmdStreamStore.keySet());
-        checkMapValues(cmdStreamStore.values());
+        checkMapKeySet(dataStreamStore.keySet());
+        checkMapValues(dataStreamStore.values());
     }
 
 
     @Test
     public void testAddAndRemoveByKey() throws Exception
     {
-        // add N different command streams
+        // add N different datastreams
         var now = TimeExtent.beginAt(Instant.now());
         int numDs = 56;
         for (int i = numDs; i < numDs*2; i++)
         {
             var sysID = bigId(i);
-            addSimpleCommandStream(sysID, "out" + (int)(Math.random()*10), now);
+            addSimpleDataStream(sysID, "out" + (int)(Math.random()*10), now);
         }
-        cmdStreamStore.commit();
+        dataStreamStore.commit();
         
-        assertEquals(numDs, cmdStreamStore.getNumRecords());
+        assertEquals(numDs, dataStreamStore.getNumRecords());
         
         int i = 0;
-        for (var id: allCmdStreams.keySet())
+        for (var id: allDataStreams.keySet())
         {
-            var ds = cmdStreamStore.remove(id);
-            checkCommandStreamEqual(allCmdStreams.get(id), ds);
+            var ds = dataStreamStore.remove(id);
+            checkDataStreamEqual(allDataStreams.get(id), ds);
             
             if (i % 5 == 0)
                 forceReadBackFromStorage();
             
             i++;
-            assertEquals(numDs-i, cmdStreamStore.getNumRecords());
+            assertEquals(numDs-i, dataStreamStore.getNumRecords());
         }
         
         // check that there is nothing left
-        assertEquals(0, cmdStreamStore.getNumRecords());
+        assertEquals(0, dataStreamStore.getNumRecords());
     }
 
 
     @Test
     public void testAddAndRemoveByFilter() throws Exception
     {
-        // add N different command streams
+        // add N different datastreams
         var idList = new ArrayList<BigId>();
         var now = TimeExtent.beginAt(Instant.now());
         int numDs = 45;
         for (int i = 1; i <= numDs; i++)
         {
             var sysID = bigId(i);
-            var key = addSimpleCommandStream(sysID, "out"+i, now);
+            var key = addSimpleDataStream(sysID, "out"+i, now);
             idList.add(key.getInternalID());
         }
-        cmdStreamStore.commit();
+        dataStreamStore.commit();
         
         int numRecords = numDs;
-        assertEquals(numRecords, cmdStreamStore.getNumRecords());
+        assertEquals(numRecords, dataStreamStore.getNumRecords());
         
         // remove some by ID
         var removedIds = new BigId[] {idList.get(3), idList.get(15), idList.get(36), idList.get(24)};
         for (var id: removedIds)
-            allCmdStreams.remove(new CommandStreamKey(id));
-        cmdStreamStore.removeEntries(new CommandStreamFilter.Builder()
+            allDataStreams.remove(new DataStreamKey(id));
+        dataStreamStore.removeEntries(new DataStreamFilter.Builder()
                 .withInternalIDs(removedIds)
                 .build());
-        checkSelectedEntries(cmdStreamStore.entrySet().stream(), allCmdStreams);
+        checkSelectedEntries(dataStreamStore.entrySet().stream(), allDataStreams);
         numRecords -= removedIds.length;
-        assertEquals(numRecords, cmdStreamStore.getNumRecords());
+        assertEquals(numRecords, dataStreamStore.getNumRecords());
         
         // remove some by name
         var removedIdsList = Arrays.asList(idList.get(4), idList.get(41), idList.get(29), idList.get(11));
         var removedNames = removedIdsList.stream()
-            .map(id -> allCmdStreams.get(new CommandStreamKey(id)).getControlInputName())
+            .map(id -> allDataStreams.get(new DataStreamKey(id)).getOutputName())
             .collect(Collectors.toList());
-        for (var id: removedIdsList)
-            allCmdStreams.remove(new CommandStreamKey(id));
-        cmdStreamStore.removeEntries(new CommandStreamFilter.Builder()
-                .withControlInputNames(removedNames)
+        for (BigId id: removedIdsList)
+            allDataStreams.remove(new DataStreamKey(id));
+        dataStreamStore.removeEntries(new DataStreamFilter.Builder()
+                .withOutputNames(removedNames)
                 .build());
-        checkSelectedEntries(cmdStreamStore.entrySet().stream(), allCmdStreams);
+        checkSelectedEntries(dataStreamStore.entrySet().stream(), allDataStreams);
         numRecords -= removedIdsList.size();
-        assertEquals(numRecords, cmdStreamStore.getNumRecords());
+        assertEquals(numRecords, dataStreamStore.getNumRecords());
         
         // remove the rest
-        cmdStreamStore.removeEntries(new CommandStreamFilter.Builder()
+        dataStreamStore.removeEntries(new DataStreamFilter.Builder()
             .build());
-        assertEquals(0, cmdStreamStore.getNumRecords());
+        assertEquals(0, dataStreamStore.getNumRecords());
     }
     
     
@@ -422,27 +432,27 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
     @SuppressWarnings("unused")
     public void testAddAndSelectCurrentVersion() throws Exception
     {
-        Stream<Entry<CommandStreamKey, ICommandStreamInfo>> resultStream;
+        Stream<Entry<DataStreamKey, IDataStreamInfo>> resultStream;
 
         var now = Instant.now().truncatedTo(ChronoUnit.SECONDS);
-        var sysID = bigId(20);
-        var ds1v0 = addSimpleCommandStream(sysID, "test1", TimeExtent.endNow(now.minusSeconds(3600)));
-        var ds1v1 = addSimpleCommandStream(sysID, "test1", TimeExtent.endNow(now.minusSeconds(1200)));
-        var ds1v2 = addSimpleCommandStream(sysID, "test1", TimeExtent.beginAt(now));
-        var ds2v0 = addSimpleCommandStream(sysID, "test2", TimeExtent.endNow(now.minusSeconds(3600)));
-        var ds2v1 = addSimpleCommandStream(sysID, "test2", TimeExtent.endNow(now));
-        var ds2v2 = addSimpleCommandStream(sysID, "test2", TimeExtent.beginAt(now.plusSeconds(600)));
-        var ds3v0 = addSimpleCommandStream(sysID, "test3", TimeExtent.endNow(now.minusSeconds(3600)));
-        var ds3v1 = addSimpleCommandStream(sysID, "test3", TimeExtent.endNow(now.minusSeconds(600)));
-        cmdStreamStore.commit();
+        var sysID = bigId(10);
+        var ds1v0 = addSimpleDataStream(sysID, "test1", TimeExtent.endNow(now.minusSeconds(3600)));
+        var ds1v1 = addSimpleDataStream(sysID, "test1", TimeExtent.endNow(now.minusSeconds(1200)));
+        var ds1v2 = addSimpleDataStream(sysID, "test1", TimeExtent.beginAt(now));
+        var ds2v0 = addSimpleDataStream(sysID, "test2", TimeExtent.endNow(now.minusSeconds(3600)));
+        var ds2v1 = addSimpleDataStream(sysID, "test2", TimeExtent.endNow(now));
+        var ds2v2 = addSimpleDataStream(sysID, "test2", TimeExtent.beginAt(now.plusSeconds(600)));
+        var ds3v0 = addSimpleDataStream(sysID, "test3", TimeExtent.endNow(now.minusSeconds(3600)));
+        var ds3v1 = addSimpleDataStream(sysID, "test3", TimeExtent.endNow(now.minusSeconds(600)));
+        dataStreamStore.commit();
         forceReadBackFromStorage();
 
         // last version of everything
-        CommandStreamFilter filter = new CommandStreamFilter.Builder()
+        DataStreamFilter filter = new DataStreamFilter.Builder()
             .withSystems(sysID)
             .withCurrentVersion()
             .build();
-        resultStream = cmdStreamStore.selectEntries(filter);
+        resultStream = dataStreamStore.selectEntries(filter);
         
         testAddAndSelectCurrentVersion_ExpectedResults();
         checkSelectedEntries(resultStream, expectedResults, filter);
@@ -459,29 +469,29 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
     @SuppressWarnings("unused")
     public void testAddAndSelectLatestValidTime() throws Exception
     {
-        Stream<Entry<CommandStreamKey, ICommandStreamInfo>> resultStream;
+        Stream<Entry<DataStreamKey, IDataStreamInfo>> resultStream;
 
         var now = Instant.now().truncatedTo(ChronoUnit.SECONDS);
         var sysID = bigId(1);
-        var ds1v0 = addSimpleCommandStream(sysID, "test1", TimeExtent.endNow(now.minusSeconds(3600)));
-        var ds1v1 = addSimpleCommandStream(sysID, "test1", TimeExtent.endNow(now.minusSeconds(1200)));
-        var ds1v2 = addSimpleCommandStream(sysID, "test1", TimeExtent.endNow(now));
-        var ds2v0 = addSimpleCommandStream(sysID, "test2", TimeExtent.endNow(now.minusSeconds(3600)));
-        var ds2v1 = addSimpleCommandStream(sysID, "test2", TimeExtent.endNow(now));
-        var ds2v2 = addSimpleCommandStream(sysID, "test2", TimeExtent.beginAt(now.plusSeconds(600)));
-        var ds3v0 = addSimpleCommandStream(sysID, "test3", TimeExtent.endNow(now.minusSeconds(3600)));
-        var ds3v1 = addSimpleCommandStream(sysID, "test3", TimeExtent.beginAt(now.minusSeconds(600)));
-        cmdStreamStore.commit();
+        var ds1v0 = addSimpleDataStream(sysID, "test1", TimeExtent.endNow(now.minusSeconds(3600)));
+        var ds1v1 = addSimpleDataStream(sysID, "test1", TimeExtent.endNow(now.minusSeconds(1200)));
+        var ds1v2 = addSimpleDataStream(sysID, "test1", TimeExtent.endNow(now));
+        var ds2v0 = addSimpleDataStream(sysID, "test2", TimeExtent.endNow(now.minusSeconds(3600)));
+        var ds2v1 = addSimpleDataStream(sysID, "test2", TimeExtent.endNow(now));
+        var ds2v2 = addSimpleDataStream(sysID, "test2", TimeExtent.beginAt(now.plusSeconds(600)));
+        var ds3v0 = addSimpleDataStream(sysID, "test3", TimeExtent.endNow(now.minusSeconds(3600)));
+        var ds3v1 = addSimpleDataStream(sysID, "test3", TimeExtent.beginAt(now.minusSeconds(600)));
+        dataStreamStore.commit();
         forceReadBackFromStorage();
 
-        // last version of everything
-        CommandStreamFilter filter = new CommandStreamFilter.Builder()
+        // latest version of everything
+        DataStreamFilter filter = new DataStreamFilter.Builder()
             .withSystems(sysID)
             .withValidTime(new TemporalFilter.Builder()
                 .withLatestTime()
                 .build())
             .build();
-        resultStream = cmdStreamStore.selectEntries(filter);
+        resultStream = dataStreamStore.selectEntries(filter);
         
         testAddAndSelectLatestValidTime_ExpectedResults();
         checkSelectedEntries(resultStream, expectedResults, filter);
@@ -498,36 +508,36 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
     @SuppressWarnings("unused")
     public void testAddAndSelectByTimeRange() throws Exception
     {
-        Stream<Entry<CommandStreamKey, ICommandStreamInfo>> resultStream;
+        Stream<Entry<DataStreamKey, IDataStreamInfo>> resultStream;
         
         var now = Instant.now().truncatedTo(ChronoUnit.SECONDS);
         var sysID1 = bigId(1);
         var sysID3 = bigId(3);
-        var ds1v0 = addSimpleCommandStream(sysID1, "out1", TimeExtent.endNow(now.minus(365, ChronoUnit.DAYS)));
-        var ds1v1 = addSimpleCommandStream(sysID1, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        var ds2v0 = addSimpleCommandStream(sysID1, "out2", TimeExtent.endNow(now.minus(520, ChronoUnit.DAYS)));
-        var ds2v1 = addSimpleCommandStream(sysID1, "out2", TimeExtent.endNow(now.minus(10, ChronoUnit.DAYS)));
-        var ds3v0 = addSimpleCommandStream(sysID1, "out3", TimeExtent.endNow(now.minus(30, ChronoUnit.DAYS)));
-        var ds3v1 = addSimpleCommandStream(sysID1, "out3", TimeExtent.endNow(now.minus(1, ChronoUnit.DAYS)));
-        var ds4v0 = addSimpleCommandStream(sysID3, "temp", TimeExtent.beginAt(now.plus(1, ChronoUnit.DAYS)));
-        var ds5v0 = addSimpleCommandStream(sysID3, "hum", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        cmdStreamStore.commit();
+        var ds1v0 = addSimpleDataStream(sysID1, "out1", TimeExtent.endNow(now.minus(365, ChronoUnit.DAYS)));
+        var ds1v1 = addSimpleDataStream(sysID1, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        var ds2v0 = addSimpleDataStream(sysID1, "out2", TimeExtent.endNow(now.minus(520, ChronoUnit.DAYS)));
+        var ds2v1 = addSimpleDataStream(sysID1, "out2", TimeExtent.endNow(now.minus(10, ChronoUnit.DAYS)));
+        var ds3v0 = addSimpleDataStream(sysID1, "out3", TimeExtent.endNow(now.minus(30, ChronoUnit.DAYS)));
+        var ds3v1 = addSimpleDataStream(sysID1, "out3", TimeExtent.endNow(now.minus(1, ChronoUnit.DAYS)));
+        var ds4v0 = addSimpleDataStream(sysID3, "temp", TimeExtent.beginAt(now.plus(1, ChronoUnit.DAYS)));
+        var ds5v0 = addSimpleDataStream(sysID3, "hum", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        dataStreamStore.commit();
         
         // select from t0 to now
-        CommandStreamFilter filter = new CommandStreamFilter.Builder()
+        DataStreamFilter filter = new DataStreamFilter.Builder()
             .withValidTimeDuring(now.minus(10, ChronoUnit.DAYS), now)
             .build();
-        resultStream = cmdStreamStore.selectEntries(filter);
+        resultStream = dataStreamStore.selectEntries(filter);
         
         testAddAndSelectByTimeRange_ExpectedResults(1);
         checkSelectedEntries(resultStream, expectedResults, filter);
                 
         // select from t0 to t1
         forceReadBackFromStorage();
-        filter = new CommandStreamFilter.Builder()
+        filter = new DataStreamFilter.Builder()
             .withValidTimeDuring(now.minus(90, ChronoUnit.DAYS), now.minus(30, ChronoUnit.DAYS))
             .build();
-        resultStream = cmdStreamStore.selectEntries(filter);
+        resultStream = dataStreamStore.selectEntries(filter);
         
         expectedResults.clear();
         testAddAndSelectByTimeRange_ExpectedResults(2);
@@ -535,11 +545,11 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
         
         // select from t0 to t1, only proc 3
         forceReadBackFromStorage();
-        filter = new CommandStreamFilter.Builder()
+        filter = new DataStreamFilter.Builder()
             .withSystems(sysID3)
             .withValidTimeDuring(now.minus(90, ChronoUnit.DAYS), now.minus(30, ChronoUnit.DAYS))
             .build();
-        resultStream = cmdStreamStore.selectEntries(filter);
+        resultStream = dataStreamStore.selectEntries(filter);
         
         expectedResults.clear();
         testAddAndSelectByTimeRange_ExpectedResults(3);
@@ -554,33 +564,33 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
             case 1: addToExpectedResults(1, 3, 4, 5, 7); break;
             case 2: addToExpectedResults(0, 1, 2, 4, 7); break;
             case 3: addToExpectedResults(7); break;
-        }
+        }        
     }
     
     
     @Test
     @SuppressWarnings("unused")
-    public void testAddAndSelectByControlInputName() throws Exception
+    public void testAddAndSelectByOutputName() throws Exception
     {
-        Stream<Entry<CommandStreamKey, ICommandStreamInfo>> resultStream;
+        Stream<Entry<DataStreamKey, IDataStreamInfo>> resultStream;
         
         var now = Instant.now().truncatedTo(ChronoUnit.SECONDS);
         var sysID1 = bigId(1);
         var sysID2 = bigId(2);
         var sysID3 = bigId(3);
-        var ds1v0 = addSimpleCommandStream(sysID1, "out1", TimeExtent.endNow(now.minus(365, ChronoUnit.DAYS)));
-        var ds1v1 = addSimpleCommandStream(sysID1, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        var ds2v0 = addSimpleCommandStream(sysID1, "out2", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        var ds3v0 = addSimpleCommandStream(sysID2, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        var ds4v0 = addSimpleCommandStream(sysID3, "temp", TimeExtent.beginAt(now.plus(1, ChronoUnit.DAYS)));
-        var ds5v0 = addSimpleCommandStream(sysID3, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        cmdStreamStore.commit();
+        var ds1v0 = addSimpleDataStream(sysID1, "out1", TimeExtent.endNow(now.minus(365, ChronoUnit.DAYS)));
+        var ds1v1 = addSimpleDataStream(sysID1, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        var ds2v0 = addSimpleDataStream(sysID1, "out2", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        var ds3v0 = addSimpleDataStream(sysID2, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        var ds4v0 = addSimpleDataStream(sysID3, "temp", TimeExtent.beginAt(now.plus(1, ChronoUnit.DAYS)));
+        var ds5v0 = addSimpleDataStream(sysID3, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        dataStreamStore.commit();
         
         // select from t0 to now
-        CommandStreamFilter filter = new CommandStreamFilter.Builder()
-            .withControlInputNames("out1")
+        DataStreamFilter filter = new DataStreamFilter.Builder()
+            .withOutputNames("out1")
             .build();
-        resultStream = cmdStreamStore.selectEntries(filter);
+        resultStream = dataStreamStore.selectEntries(filter);
         
         testAddAndSelectByOutputName_ExpectedResults();
         checkSelectedEntries(resultStream, expectedResults, filter);
@@ -597,25 +607,25 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
     @SuppressWarnings("unused")
     public void testAddAndSelectBySystemID() throws Exception
     {
-        Stream<Entry<CommandStreamKey, ICommandStreamInfo>> resultStream;
+        Stream<Entry<DataStreamKey, IDataStreamInfo>> resultStream;
         
         var now = Instant.now().truncatedTo(ChronoUnit.SECONDS);
         var sysID1 = bigId(1);
         var sysID2 = bigId(2);
         var sysID3 = bigId(3);
-        var ds1v0 = addSimpleCommandStream(sysID1, "out1", TimeExtent.endNow(now.minus(365, ChronoUnit.DAYS)));
-        var ds1v1 = addSimpleCommandStream(sysID1, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        var ds2v0 = addSimpleCommandStream(sysID1, "out2", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        var ds3v0 = addSimpleCommandStream(sysID2, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        var ds4v0 = addSimpleCommandStream(sysID3, "temp", TimeExtent.beginAt(now.plus(1, ChronoUnit.DAYS)));
-        var ds5v0 = addSimpleCommandStream(sysID3, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        cmdStreamStore.commit();
+        var ds1v0 = addSimpleDataStream(sysID1, "out1", TimeExtent.endNow(now.minus(365, ChronoUnit.DAYS)));
+        var ds1v1 = addSimpleDataStream(sysID1, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        var ds2v0 = addSimpleDataStream(sysID1, "out2", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        var ds3v0 = addSimpleDataStream(sysID2, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        var ds4v0 = addSimpleDataStream(sysID3, "temp", TimeExtent.beginAt(now.plus(1, ChronoUnit.DAYS)));
+        var ds5v0 = addSimpleDataStream(sysID3, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        dataStreamStore.commit();
         
         // select from t0 to now
-        CommandStreamFilter filter = new CommandStreamFilter.Builder()
+        DataStreamFilter filter = new DataStreamFilter.Builder()
             .withSystems(sysID2, sysID3)
             .build();
-        resultStream = cmdStreamStore.selectEntries(filter);
+        resultStream = dataStreamStore.selectEntries(filter);
         
         testAddAndSelectBySystemID_ExpectedResults();
         checkSelectedEntries(resultStream, expectedResults, filter);
@@ -632,68 +642,68 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
     @SuppressWarnings("unused")
     public void testAddAndSelectByKeywords() throws Exception
     {
-        Stream<Entry<CommandStreamKey, ICommandStreamInfo>> resultStream;
-        CommandStreamFilter filter;
+        Stream<Entry<DataStreamKey, IDataStreamInfo>> resultStream;
+        DataStreamFilter filter;
         
         var now = Instant.now().truncatedTo(ChronoUnit.SECONDS);
         var sysID1 = bigId(1);
         var sysID2 = bigId(2);
         var sysID3 = bigId(3);
-        var ds1v0 = addSimpleCommandStream(sysID1, "out1", "Stationary weather data", TimeExtent.endNow(now.minus(365, ChronoUnit.DAYS)));
-        var ds1v1 = addSimpleCommandStream(sysID1, "out1", "Stationary weather data", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        var ds2v0 = addSimpleCommandStream(sysID1, "out2", "Traffic video stream", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        var ds3v0 = addSimpleCommandStream(sysID2, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        var ds4v0 = addSimpleCommandStream(sysID3, "temp", "Air temperature", TimeExtent.beginAt(now.plus(1, ChronoUnit.DAYS)));
-        var ds5v0 = addSimpleCommandStream(sysID3, "out1", "Air pressure", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
-        cmdStreamStore.commit();
+        var ds1v0 = addSimpleDataStream(sysID1, "out1", "Stationary weather data", TimeExtent.endNow(now.minus(365, ChronoUnit.DAYS)));
+        var ds1v1 = addSimpleDataStream(sysID1, "out1", "Stationary weather data", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        var ds2v0 = addSimpleDataStream(sysID1, "out2", "Traffic video stream", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        var ds3v0 = addSimpleDataStream(sysID2, "out1", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        var ds4v0 = addSimpleDataStream(sysID3, "temp", "Air temperature", TimeExtent.beginAt(now.plus(1, ChronoUnit.DAYS)));
+        var ds5v0 = addSimpleDataStream(sysID3, "out1", "Air pressure", TimeExtent.endNow(now.minus(60, ChronoUnit.DAYS)));
+        dataStreamStore.commit();
         
         // select with one keyword
-        filter = new CommandStreamFilter.Builder()
+        filter = new DataStreamFilter.Builder()
             .withKeywords("air")
             .build();
-        resultStream = cmdStreamStore.selectEntries(filter);
+        resultStream = dataStreamStore.selectEntries(filter);
         
         expectedResults.clear();
         testAddAndSelectByKeywords_ExpectedResults(1);
         checkSelectedEntries(resultStream, expectedResults, filter);
         
         // select with 2 keywords
-        filter = new CommandStreamFilter.Builder()
+        filter = new DataStreamFilter.Builder()
             .withKeywords("air", "weather")
             .build();
-        resultStream = cmdStreamStore.selectEntries(filter);
+        resultStream = dataStreamStore.selectEntries(filter);
         
         expectedResults.clear();
         testAddAndSelectByKeywords_ExpectedResults(2);
         checkSelectedEntries(resultStream, expectedResults, filter);
         
         // select with 2 keywords
-        filter = new CommandStreamFilter.Builder()
+        filter = new DataStreamFilter.Builder()
             .withKeywords("air", "video")
             .build();
-        resultStream = cmdStreamStore.selectEntries(filter);
+        resultStream = dataStreamStore.selectEntries(filter);
         
         expectedResults.clear();
         testAddAndSelectByKeywords_ExpectedResults(3);
         checkSelectedEntries(resultStream, expectedResults, filter);
         
         // select with system and keywords (partial words)
-        filter = new CommandStreamFilter.Builder()
+        filter = new DataStreamFilter.Builder()
             .withSystems(sysID3)
             .withKeywords("weather", "temp")
             .build();
-        resultStream = cmdStreamStore.selectEntries(filter);
+        resultStream = dataStreamStore.selectEntries(filter);
         
         expectedResults.clear();
         testAddAndSelectByKeywords_ExpectedResults(4);
         checkSelectedEntries(resultStream, expectedResults, filter);
         
         // select unknown keywords
-        filter = new CommandStreamFilter.Builder()
+        filter = new DataStreamFilter.Builder()
             .withSystems(sysID3)
             .withKeywords("lidar", "humidity")
             .build();
-        resultStream = cmdStreamStore.selectEntries(filter);
+        resultStream = dataStreamStore.selectEntries(filter);
         
         expectedResults.clear();
         testAddAndSelectByKeywords_ExpectedResults(5);
@@ -709,18 +719,18 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
             case 2: addToExpectedResults(0, 1, 4, 5); break;
             case 3: addToExpectedResults(2, 4, 5); break;
             case 4: addToExpectedResults(4); break;
-        }
+        }        
     }
     
     
     @Test(expected = DataStoreException.class)
-    public void testErrorAddWithExistingInput() throws Exception
+    public void testErrorAddWithExistingOutput() throws Exception
     {
         var now = Instant.now().truncatedTo(ChronoUnit.SECONDS);
         var sysID1 = bigId(1);
-        addSimpleCommandStream(sysID1, "cmd1", TimeExtent.beginAt(now));
-        addSimpleCommandStream(sysID1, "cmd1", TimeExtent.beginAt(now));
-        cmdStreamStore.commit();
+        addSimpleDataStream(sysID1, "out1", TimeExtent.beginAt(now));
+        addSimpleDataStream(sysID1, "out1", TimeExtent.beginAt(now));
+        dataStreamStore.commit();
     }
     
     
@@ -729,7 +739,7 @@ public abstract class AbstractTestCommandStreamStore<StoreType extends ICommandS
     {
         try
         {
-            cmdStreamStore.selectEntries(new CommandStreamFilter.Builder()
+            dataStreamStore.selectEntries(new DataStreamFilter.Builder()
                 .withSystems()
                     .withKeywords("thermometer")
                     .done()
