+++ sensorhub-core/src/testFixtures/java/org/sensorhub/impl/event/TestEvent.java
