+++ sensorhub-core/src/testFixtures/java/org/sensorhub/impl/module/AsyncModule.java
@@ -14,15 +14,16 @@ Copyright (C) 2012-2016 Sensia Software LLC. All Rights Reserved.
 
 package org.sensorhub.impl.module;
 
-import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import org.sensorhub.api.event.Event;
 import org.sensorhub.api.common.SensorHubException;
+import org.sensorhub.api.event.Event;
 import org.sensorhub.api.module.IModule;
 import org.sensorhub.api.module.ModuleEvent;
 import org.sensorhub.api.module.ModuleEvent.ModuleState;
 import org.sensorhub.utils.MsgUtils;
+
+import java.util.concurrent.Callable;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 
 public class AsyncModule extends AbstractModule<AsyncModuleConfig>
