+++ sensorhub-core/src/testFixtures/java/org/sensorhub/impl/module/DummyModule.java
