+++ sensorhub-core/src/testFixtures/java/org/sensorhub/impl/module/MyConfig1.java
