+++ sensorhub-core/src/testFixtures/java/org/sensorhub/impl/module/MyConfig2.java
