+++ sensorhub-core/src/testFixtures/java/org/sensorhub/impl/processing/TestLinearEquationProcessModule.java
