+++ sensorhub-core/src/testFixtures/java/org/sensorhub/impl/sensor/FakeSensorData.java
@@ -14,18 +14,19 @@ Copyright (C) 2012-2015 Sensia Software LLC. All Rights Reserved.
 
 package org.sensorhub.impl.sensor;
 
+import net.opengis.swe.v20.DataBlock;
+import net.opengis.swe.v20.DataComponent;
+import net.opengis.swe.v20.DataEncoding;
+import org.sensorhub.api.data.DataEvent;
+import org.sensorhub.api.data.IDataProducer;
+import org.vast.data.DataBlockDouble;
+import org.vast.data.TextEncodingImpl;
+import org.vast.swe.SWEHelper;
+
 import java.time.Instant;
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.concurrent.CompletableFuture;
-import net.opengis.swe.v20.DataComponent;
-import net.opengis.swe.v20.DataEncoding;
-import net.opengis.swe.v20.DataBlock;
-import org.sensorhub.api.data.IDataProducer;
-import org.sensorhub.api.data.DataEvent;
-import org.vast.data.DataBlockDouble;
-import org.vast.data.TextEncodingImpl;
-import org.vast.swe.SWEHelper;
 
 
 /**
