+++ sensorhub-core/src/testFixtures/java/org/sensorhub/impl/sensor/FakeSensorNetWithMembers.java
@@ -14,14 +14,7 @@ Copyright (C) 2012-2015 Sensia Software LLC. All Rights Reserved.
 
 package org.sensorhub.impl.sensor;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.concurrent.CompletableFuture;
-import java.util.function.Consumer;
-import javax.xml.namespace.QName;
+import com.google.common.collect.ImmutableMap;
 import net.opengis.gml.v32.AbstractFeature;
 import net.opengis.gml.v32.Point;
 import net.opengis.gml.v32.impl.GMLFactory;
@@ -38,7 +31,11 @@ import org.sensorhub.impl.event.EventSourceInfo;
 import org.vast.ogc.gml.GenericFeatureImpl;
 import org.vast.ogc.gml.IFeature;
 import org.vast.sensorML.SMLFactory;
-import com.google.common.collect.ImmutableMap;
+
+import javax.xml.namespace.QName;
+import java.util.*;
+import java.util.concurrent.CompletableFuture;
+import java.util.function.Consumer;
 
 
 public class FakeSensorNetWithMembers extends FakeSensor implements ISystemGroupDriver<IDataProducer>
