+++ sensorhub-core/src/testFixtures/java/org/sensorhub/impl/sensor/FakeSensorWithPos.java
