+++ sensorhub-core/src/testFixtures/java/org/sensorhub/impl/sensor/IFakeSensorOutput.java
